### 2023-04-18 #5384 @kthyng
Hi @Zeitsperre and thanks for your submission! We are happy to publish this after its review with pyOpenSci. However, we wait until after that point. It looks like it is still under review there? If so, I could add the "paused" label to this until it is finished, then pick it back up to go through the final steps. Sound good?
