### 2023-04-27 #5415 @editorialbot
[REVIEW]: xclim: xarray-based climate data analytics
**Submitting author:** @Zeitsperre (<a href="http://orcid.org/0000-0001-5393-8359">Trevor James Smith</a>)
**Repository:** https://github.com/Ouranosinc/xclim
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v0.43.0
