### 2023-04-27 #5415 @kthyng
Paper looks good ✅

@Zeitsperre Do you have a Zenodo archive associated with your pyOpenSci review?
