### 2023-05-01 #5415 @Zeitsperre
Hi @kthyng

Please excuse the late response (I was on vacation).

We have a permanent Zenodo DOI, but the last version does not yet include the pyOpenSci-approved build version (we are releasing a new version on 5 May). When released, it should be available automatically and immediately here: https://zenodo.org/record/7796479
