### 2023-05-08 #5415 @kthyng
Hi @Zeitsperre! We should associate the DOI associated with the pyOpenSci review with the JOSS paper, so it looks like we need to wait for that? Also we request for JOSS submissions that the metadata for that archive be altered so that the title and author list exactly match the JOSS paper (looks more polished that way) — could you do that for the upcoming release?
