### 2023-05-09 #5415 @Zeitsperre
@kthyng

We've just released a new version of xclim on PyPI (conda-forge later today) and have synchronized the authorship for Zenodo and the paper submission. Let me know if there's anything else that needs doing on my end.

Thanks!
