### 2023-05-15 #5415 @Zeitsperre
Hi @kthyng,

I don't think I have control over the title and information generated by @editorialbot, but yes, we would be submitting v0.43.0. I imagine that for the DOI, you need the static DOI to v0.43.0? In that case, yes, 10.5281/zenodo.7916446 looks good.

It's unfortunate, but the Zenodo and paper references were only synchronized in the branch containing the paper (https://github.com/Ouranosinc/xclim/tree/joss_submission). I'm not sure if this is a blocker for the review moving forward.
