### 2023-05-15 #5415 @kthyng
@Zeitsperre I'm not sure we're talking about the same thing — let me try again. As you said, yes we want to have a specific match from your code as a release to the JOSS paper since of course the code is changing over time but a specific version was reviewed. But for that specific Zenodo release we prefer to have the title in the metadata match your JOSS paper and your author list match the JOSS paper, if possible (not required). I believe you can edit the metadata by hand for a specific version even if normally it is done automatically. So I would be looking for the author list and title to match your JOSS paper for this archive version: https://zenodo.org/record/7916446#.ZGJMNuzMLlw

But, if it isn't possible or you don't want to, let me know. And sorry if you are addressing this but I'm just not understanding!
