### 2023-05-15 #5415 @Zeitsperre
Ah, I think I understand now. We release fairly regularly, so we tend to simply name our releases with the version numbers, but I see now that we can rename v0.43.0 on Zenodo with the title as written in the paper. Thanks for your patience!

For the author list, it would be great to have the author order as listed in the paper. I'll see if I can update this list momentarily.
