### 2023-05-15 #5415 @Zeitsperre
@kthyng I just checked again, and I think these are in sync.

From the paper:
```markdown
 - name: Pascal Bourgault
 - name: David Huard
 - name: Trevor James Smith
 - name: Travis Logan
 - name: Abel Aoun
 - name: Juliette Lavoie
 - name: Éric Dupuis
 - name: Gabriel Rondeau-Genessse
 - name: Raquel Alegre
 - name: Clair Barnes
 - name: Alexis Beaupré Laperrière
 - name: Sébastien Biner
 - name: David Caron
 - name: Carsten Ehbrecht
 - name: Jeremy Fyke
 - name: Tom Keel
 - name: Marie-Pier Labonté
 - name: Ludwig Lierhammer
 - name: Jwen-Fai Low
 - name: Jamie Quinn
 - name: Philippe Roy
 - name: Dougie Squire
 - name: Ag Stephens
 - name: Maliko Tanguy
 - name: Christopher Whelan
```

On Zenodo:
![image](https://github.com/openjournals/joss-reviews/assets/10819524/d41ad24b-8576-4543-aeb1-33fc1adcc79e)

For each version, we also list the names of the contributors that added code specifically for that release, so this list will be much shorter, i.e.:

![image](https://github.com/openjournals/joss-reviews/assets/10819524/a8e4a855-389c-48d7-8465-914af92b5097)

Could that be the source of confusion?
