### 2023-05-15 #5415 @Zeitsperre
The paper is currently on a branch, but we run the paper generating action: https://github.com/Ouranosinc/xclim/actions/runs/4929890050?pr=250
