### 2023-05-15 #5415 @kthyng
Duh, I was looking at an old version of the paper for the author list, sorry for that confusion. Your branch is set so that part should be ok, but I'm not sure why your building of the paper is not hitting an error. Still, you can see [here](https://github.com/openjournals/joss-papers/actions/runs/4982835868/jobs/8919067529) what it is complaining about — not everyone has an affiliation but needs to, I guess.
