### 2023-05-15 #5415 @Zeitsperre
Ah, I see. I can fill in most of them, but there are a few authors who I have no idea (and no easy way of contacting them). Is there an acceptable way of indicating "none" for those cases?
