### 2023-05-15 #5415 @kthyng
You can see an example [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). It would be best to get something reasonable for everyone, but if you absolutely can't you could use the "independent researcher" one, something like that. But it is something that will still come through to the final publication as a listing.
